requirements.txt
absl-py
tensorflow>=2.0.0
transformers
torch
seqeval
pytorch-crf
